KlayGE: Core: Fix compiling error under x64. KlayGE: Tools: Add GL 4.4 support in GLCompatibility.

+++ KlayGE/Core/Src/Render/ParticleSystem.cpp
@@ -749,7 +749,7 @@ namespace KlayGE
 			}
 		}
 
-		num_active_particles_ = active_particles.size();
+		num_active_particles_ = static_cast<uint32_t>(active_particles.size());
 
 		if (!active_particles.empty())
 		{
